feat(205-01): extract useEntitySave generic hook, convert save hooks to thin wrappers
Create use-entity-save.ts with parameterized save/discard logic that
accepts an EntitySaveConfig. Convert use-agent-save, use-skill-save,
and use-rule-save to thin wrappers delegating to the generic hook.
Zero consumer changes required.
